backend/requirements-dev.txt
watchfiles==1.1.0
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5